Create a new voice package to support external voice connections (#44)
