+++ client.go
@@ -2,6 +2,7 @@ package harmony
 
 import (
 	"net/http"
+	"os"
 	"sync"
 	"time"
 
@@ -10,6 +11,7 @@ import (
 	"github.com/skwair/harmony/internal/payload"
 	"github.com/skwair/harmony/internal/rate"
 	"github.com/skwair/harmony/log"
+	"github.com/skwair/harmony/voice"
 )
 
 const (
@@ -110,6 +112,10 @@ type Client struct {
 	withStateTracking bool
 	State             *State
 
+	// voice connections that were established by
+	// this client.
+	voiceConnections map[string]*voice.Connection
+
 	logger log.Logger
 }
 
@@ -131,7 +137,8 @@ func NewClient(token string, opts ...ClientOption) (*Client, error) {
 		handlers:           make(map[string]handler),
 		backoff:            defaultBackoff,
 		withStateTracking:  true,
-		logger:             log.NewStd(log.LevelError),
+		voiceConnections:   make(map[string]*voice.Connection),
+		logger:             log.NewStd(os.Stderr, log.LevelError),
 	}
 
 	for _, opt := range opts {
